Improve handling of null when resolving nested object
If a property inside is null, simply keep it as null instead of having
the automatically created object resolver call its nested resolvers
with null.

Example:

With a resolver structure like:

     {
       page: {
         type: resolver(...)
       }
     }

For an object like

     {
       page: null
     }

it is an error to call the resolver for `type` with null as
`props`. Instead the implicit object resolver for `page` can simply
short circuit and resolve to `null`.